% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{compute_average_weights_ivlike}
\alias{compute_average_weights_ivlike}
\title{Compute Average Weights for IV-like Parameters}
\usage{
compute_average_weights_ivlike(ivlike, dgp)
}
\arguments{
\item{ivlike}{A list containing the name of the IV-like parameter and its function 's'.}

\item{dgp}{A DGP object containing the data generating process.}
}
\value{
A data frame containing the average weights.
}
\description{
This function computes the average weights for IV-like parameters.
}
