% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{constantspline_basis}
\alias{constantspline_basis}
\title{Create a constant spline basis}
\usage{
constantspline_basis(knots)
}
\arguments{
\item{knots}{A vector of knot points, must include 0 and 1.}
}
\value{
An MTRBasis list for the constant spline basis.
}
\description{
This function generates a constant spline basis based on the given knots.
}
