% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{create_dgp}
\alias{create_dgp}
\title{Create DGP Object for MST2018 Model}
\usage{
create_dgp(suppZ, densZ, pscoreZ, basis0, theta0, basis1, theta1)
}
\value{
A new DGP object.
}
\description{
This function creates a DGP object based on inputs.
}
\section{Fields}{

\describe{
\item{\code{suppZ}}{Support of Z.}

\item{\code{densZ}}{Density of Z.}

\item{\code{pscore}}{Propensity score.}

\item{\code{mtrs}}{List of Marginal Treatment Response (MTR) functions.}
}}

