% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{create_plot}
\alias{create_plot}
\title{Create a ggplot based on various conditions and data}
\usage{
create_plot(dgp, d, upper, weights_tp, weights_ivlike, bounds, basis, title)
}
\arguments{
\item{dgp}{Data Generating Process object containing necessary data and parameters.}

\item{d}{The treatment indicator (0 or 1).}

\item{upper}{Logical indicating whether to use upper bounds.}

\item{weights_tp}{List of weights for the treatment parameters.}

\item{weights_ivlike}{List of weights for IV-like estimators.}

\item{bounds}{Object containing upper and lower bound solutions.}

\item{basis}{The type of basis to use ("constant_spline", "weighted_bernstein", or "dgp_mtr").}

\item{title}{The title for the plot.}
}
\value{
A ggplot object.
}
\description{
This function generates a ggplot object based on the provided data, conditions, and parameters.
The function supports different basis types and allows for upper and lower bounds.
}
