% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{DGP}
\alias{DGP}
\title{DGP R6 Class}
\description{
An R6 class for representing a Data Generating Process (DGP).
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{suppZ}}{Support of Z.}

\item{\code{densZ}}{Density of Z.}

\item{\code{pscore}}{Propensity score.}

\item{\code{mtrs}}{List of Marginal Treatment Response (MTR) functions.
Initialize DGP object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DGP-new}{\code{DGP$new()}}
\item \href{#method-DGP-find_pscore}{\code{DGP$find_pscore()}}
\item \href{#method-DGP-find_density}{\code{DGP$find_density()}}
\item \href{#method-DGP-clone}{\code{DGP$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DGP-new"></a>}}
\if{latex}{\out{\hypertarget{method-DGP-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DGP$new(suppZ, densZ, pscore, mtrs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{suppZ}}{Support of Z.}

\item{\code{densZ}}{Density of Z.}

\item{\code{pscore}}{Propensity score.}

\item{\code{mtrs}}{List of MTR functions.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new DGP object.
Find Propensity Score for Given Z
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DGP-find_pscore"></a>}}
\if{latex}{\out{\hypertarget{method-DGP-find_pscore}{}}}
\subsection{Method \code{find_pscore()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DGP$find_pscore(z)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{z}}{Value of Z.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Propensity score corresponding to Z.
Find Density for Given Z
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DGP-find_density"></a>}}
\if{latex}{\out{\hypertarget{method-DGP-find_density}{}}}
\subsection{Method \code{find_density()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DGP$find_density(z)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{z}}{Value of Z.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Density corresponding to Z.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DGP-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DGP-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DGP$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
