% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtr.R
\name{evaluate_mtr}
\alias{evaluate_mtr}
\title{Evaluate the MTR (Marginal Treatment Response) at given points}
\usage{
evaluate_mtr(mtr, ev)
}
\arguments{
\item{mtr}{An MTR object containing basis functions and theta.}

\item{ev}{A data frame containing the evaluation points (z and u).}
}
\value{
A numeric vector containing the evaluated MTR values.
}
\description{
Evaluate the MTR (Marginal Treatment Response) at given points
}
