% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtr.R
\name{evaluate_mtr_tuple}
\alias{evaluate_mtr_tuple}
\title{Evaluate the MTR (Marginal Treatment Response) tuple at given points}
\usage{
evaluate_mtr_tuple(mtrs, ev)
}
\arguments{
\item{mtrs}{A list of MTR objects for each treatment level (0 and 1).}

\item{ev}{A data frame containing the evaluation points (z and u) and treatment assignment (d).}
}
\value{
A numeric vector containing the evaluated MTR values.
}
\description{
Evaluate the MTR (Marginal Treatment Response) tuple at given points
}
