% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{integrate_bernstein_polynomial}
\alias{integrate_bernstein_polynomial}
\title{Integrate kth Bernstein basis polynomial of degree K}
\usage{
integrate_bernstein_polynomial(lower_pt, upper_pt, k, K)
}
\arguments{
\item{lower_pt}{The lower limit of integration.}

\item{upper_pt}{The upper limit of integration.}

\item{k}{The index of the Bernstein basis polynomial.}

\item{K}{The degree of the Bernstein basis.}
}
\value{
The integral value.
}
\description{
This function calculates the integral of the kth Bernstein basis polynomial of degree K between lower_pt and upper_pt.
}
