% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivlike.R
\name{ivslope}
\alias{ivslope}
\title{Compute IV Slope}
\usage{
ivslope(dgp)
}
\arguments{
\item{dgp}{Data Generating Process object.}
}
\value{
A list containing the name and s function.
}
\description{
Compute IV Slope
}
