% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivlike.R
\name{make_slist}
\alias{make_slist}
\title{Create Saturated IV-like List}
\usage{
make_slist(dgp)
}
\arguments{
\item{dgp}{Data Generating Process object.}
}
\value{
A list containing the name, s functions, and legend titles.
}
\description{
Create Saturated IV-like List
}
