% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{weighted_bernstein_basis}
\alias{weighted_bernstein_basis}
\title{Weighted Kth degree Bernstein basis polynomial}
\usage{
weighted_bernstein_basis(u, K, weights)
}
\arguments{
\item{u}{The point(s) at which to evaluate the polynomial.}

\item{K}{The degree of the Bernstein basis.}

\item{weights}{A vector of weights for each basis polynomial.}
}
\value{
The weighted sum of the Bernstein basis polynomials.
}
\description{
This function calculates a weighted sum of Bernstein basis polynomials of degree K at a given point u.
}
